Removed ScalaTemplateParser dependencies on scala-ide, ScalaTempalateCompiler, 2.10.x
ScalaTemplateParser depended on:
* The TemplateTree classes in ScalaTempalteCompiler ==> These have been refactored to a separate file.
* HasLogger ==> Logging has simply been removed
* String interpolation ==> replaced with simple string concatenation [to make compatible with scala 2.9]
* ClassTags/TypeTags ==> replaced with deprecated Manifest [to make compatible with scala 2.9]

These changes are in preparation for when ScalaTemplateParser is split off into it's own project.
